offline login

+++ application/app/src/main/java/com/application/keykeeper/StorageFragment.kt
@@ -22,6 +22,7 @@ class StorageFragment: Fragment() {
     private lateinit var adapter: CredentialsAdapter
     private lateinit var toolbar: Toolbar
     private lateinit var toolbarAddItemButton: ImageButton
+    private var offlineMode: Boolean = false
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
@@ -49,6 +50,10 @@ class StorageFragment: Fragment() {
 
         // Allow the search bar to be opened from anywhere in its area, and not only the icon.
         searchView.setOnClickListener { searchView.isIconified = false }
+
+        // Hide the add item button if in offline mode
+        offlineMode = Model.Storage.inOfflineMode(context!!)
+        toolbarAddItemButton.visibility = if (offlineMode) View.INVISIBLE else View.VISIBLE
 
         // Setup toolbar button
         toolbarAddItemButton.setOnClickListener {
@@ -108,6 +113,11 @@ class StorageFragment: Fragment() {
         val editButton = view.findViewById<Button>(R.id.storage_item_popup_edit_button)
         val deleteButton = view.findViewById<Button>(R.id.storage_item_popup_delete_button)
 
+        // Grey out buttons if in offline mode
+        if (offlineMode) {
+            editButton.alpha = .5f
+            deleteButton.alpha = .5f
+        }
 
         // Display hidden password
         passwordLabel.text = item.password.replace(".".toRegex(), "*")
@@ -126,31 +136,35 @@ class StorageFragment: Fragment() {
 
         // Open edit popup when clicked
         editButton.setOnClickListener {
-            Utils.hideStatusMessage(statusMessage)
-            openEditItemPopup(adapter, index).setOnDismissListener {
-                // Dismiss this if still open when the editor closes, possibly contains invalid data
-                // TODO: Fix overlaying backgrounds
-                window.dismiss()
+            if (!offlineMode) {
+                Utils.hideStatusMessage(statusMessage)
+                openEditItemPopup(adapter, index).setOnDismissListener {
+                    // Dismiss this if still open when the editor closes, possibly contains invalid data
+                    // TODO: Fix overlaying backgrounds
+                    window.dismiss()
+                }
             }
         }
 
         // Delete confirmation
         deleteButton.setOnClickListener {
-            Utils.hideStatusMessage(statusMessage)
-            if (deleteButton.tag == "delete") {
-                deleteButton.tag = "confirm"
-                deleteButton.setText(R.string.storage_popup_confirm_delete)
-                deleteButton.setBackgroundColor(context.resources.getColor(R.color.light_dangerous))
-            } else {
-                Model.Communication.deleteItem(item) { successful, message ->
-                    if (successful) {
-                        deleteButton.tag = "delete"
-                        deleteButton.setText(R.string.storage_popup_delete)
-                        deleteButton.setBackgroundColor(context.resources.getColor(R.color.dangerous))
-                        fetchAndUpdateListView()
-                        window.dismiss()
-                    } else {
-                        Utils.showStatusMessage(statusMessage, message, true)
+            if (!offlineMode) {
+                Utils.hideStatusMessage(statusMessage)
+                if (deleteButton.tag == "delete") {
+                    deleteButton.tag = "confirm"
+                    deleteButton.setText(R.string.storage_popup_confirm_delete)
+                    deleteButton.setBackgroundColor(context.resources.getColor(R.color.light_dangerous))
+                } else {
+                    Model.Communication.deleteItem(item) { successful, message ->
+                        if (successful) {
+                            deleteButton.tag = "delete"
+                            deleteButton.setText(R.string.storage_popup_delete)
+                            deleteButton.setBackgroundColor(context.resources.getColor(R.color.dangerous))
+                            fetchAndUpdateListView()
+                            window.dismiss()
+                        } else {
+                            Utils.showStatusMessage(statusMessage, message, true)
+                        }
                     }
                 }
             }
